combine "instnance" and "module" load code
