--- lib/Module.js
@@ -1,32 +1,0 @@
-'use strict'
-
-module.exports = function (scope, module_iri, session, callback) {
-
-    let module = scope.cache.get(module_iri);
-    if (module) {
-        module.ready ? callback(null, module.exports) : module.queue.push(callback);
-        return;
-    }
-
-    module = {queue: [callback]};
-    scope.cache.set(module_iri, module);
-    scope.mod(module_iri, session, (error, exports) => {
-
-        if (error) {
-            scope.cache.del(module_iri);
-            return loaded(error, module);
-        }
-
-        module.exports = exports;
-        module.ready = true;
-        scope.cache.set(module_iri, module);
-        loaded(null, module);
-    });
-
-    return module;
-};
-
-function loaded (error, module) {
-    module.queue.forEach(callback => callback(error, module.exports));
-    module.queue = [];
-}
